Fix NPE and delete shops when resident purges

+++ src/main/java/dev/warriorrr/quickshopremover/TownyListener.java
@@ -1,8 +1,10 @@
 package dev.warriorrr.quickshopremover;
 
+import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
+import com.palmergames.bukkit.towny.event.DeletePlayerEvent;
 import com.palmergames.bukkit.towny.event.town.TownUnclaimEvent;
 
 import org.bukkit.Chunk;
@@ -25,8 +27,18 @@ public class TownyListener implements Listener {
     public void onTownUnclaim(TownUnclaimEvent event) {
         Chunk unclaimedChunk = event.getWorldCoord().getBukkitWorld().getChunkAt(event.getWorldCoord().getX(), event.getWorldCoord().getZ());
         Map <Location,Shop> shopMap = QuickShopAPI.getShopAPI().getShop(unclaimedChunk);
-        for (Entry<Location, Shop> shop : shopMap.entrySet()) {
-            shop.getValue().delete();
+        if (shopMap.size() > 0) {
+            for (Entry<Location, Shop> shop : shopMap.entrySet())
+                shop.getValue().delete();
+        }
+    }
+
+    @EventHandler
+    public void onResidentDelete(DeletePlayerEvent event) {
+        List<Shop> ownedShops = QuickShopAPI.getShopAPI().getPlayerAllShops(event.getPlayerUUID());
+        if (ownedShops.size() > 0) {
+            for (Shop shop : ownedShops)
+                shop.delete();
         }
     }
 }
